requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
blinker==1.7.0
boto3==1.28.77
botocore==1.31.77
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
constantly==23.10.4
cryptography==41.0.5
cs==3.2.0
cssselect==1.2.0
Faker==20.0.0
filelock==3.13.1
Flask==3.0.0
frozenlist==1.4.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
lxml==4.9.3
MarkupSafe==2.1.3
multidict==6.0.4
nltk==3.8.1
numpy==1.26.1
openai==0.28.1
packaging==23.2
parsel==1.8.1
Protego==0.3.0
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
PyMySQL==1.1.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
queuelib==1.6.2
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
s3transfer==0.7.0
scikit-learn==1.3.2
scipy==1.11.3
Scrapy==2.11.0
service-identity==23.1.0
six==1.16.0
sklearn==0.0
soupsieve==2.5
SQLAlchemy==2.0.23
text-unidecode==1.3
threadpoolctl==3.2.0
tldextract==5.1.0
tqdm==4.66.1
Twisted==22.10.0
typing_extensions==4.8.0
urllib3==2.0.7
w3lib==2.1.2
Werkzeug==3.0.1
yarl==1.9.2
zope.interface==6.1
